+++ CHANGES.md
@@ -1,4 +1,4 @@
-x.x
+0.5
 - Made widget form smaller
 - Added widget output filter and default widget output
 - Added template loader